Back-End/requirements.txt
# Django
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5

# API Documentation
drf-spectacular>=0.27.0
drf-spectacular-sidecar>=2024.1.1

# Base de datos (opcional para producción)
# psycopg2-binary>=2.9.9  # Para PostgreSQL

# Utilidades
python-dotenv>=1.0.0
